default to first feature preview in query
